backend/requirements.txt
# Updated requirements.txt with older azure-ai-openai version
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
email-validator==2.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
httpx==0.24.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
aiohttp==3.8.4
aiocron==1.8

# Azure services
azure-identity==1.13.0
azure-search-documents==11.4.0
azure-ai-formrecognizer==3.2.0
azure-ai-textanalytics==5.2.0
azure-ai-openai==0.7.0  # Changed to version less than 1.0.0
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.7.1